% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accept.R
\name{accept}
\alias{accept}
\title{Accept ---- helper function ------------------------------------------------}
\usage{
accept(
  z,
  .odds_ratio,
  .m,
  .n,
  .df,
  .alpha,
  .precision,
  .method,
  .maze,
  .zoom_iter
)
}
\arguments{
\item{z}{Vector containing (u, t).}

\item{.odds_ratio}{The null hypothesis odds ratio being tested. No default.}

\item{.m}{Integer input responses and sample sizes. Tests u/m versus v/n. No default.}

\item{.n}{Integer input responses and sample sizes. Tests u/m versus v/n. No default.}

\item{.df}{Testing frame (data frame) generated as part as construct_test_frame().}

\item{.alpha}{The nominal significance level α. Defaults to 0.05.}

\item{.precision}{Defines the precision by which confidence limits, p-values, and size is determined. Defaults to 1E-03.}

\item{.method}{Defines the numerical method used to find the optimum nuisance parameter (maximising actual size) of the test. The default is "zoom", with the second option being "trust" (this uses the trust() function from the trust region package).}

\item{.maze}{Number of points at each iteration to select the nuisance parameter with maximum size from.}

\item{.zoom_iter}{Number of iterations to zoom in with (in the "zoom" method).}
}
\description{
Determine if H : theta = theta0 is accepted based on MFET. Used within
modified fisher exact test function and not exported.
}
\keyword{accept}
\keyword{exact}
\keyword{fisher}
\keyword{hypothesis}
\keyword{modified}
\keyword{reject}
\keyword{test}
