% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_exp_value.R
\name{calc_expected_value}
\alias{calc_expected_value}
\title{Calculate expected OR value ---- helper function ---------------------------}
\usage{
calc_expected_value(.odds_ratio, .m, .n, .t, .precision)
}
\arguments{
\item{.odds_ratio}{The null hypothesis odds ratio being tested. No default.}

\item{.m}{Integer input responses and sample sizes. Tests u/m versus v/n. No default.}

\item{.n}{Integer input responses and sample sizes. Tests u/m versus v/n. No default.}

\item{.t}{Integer input responses and sample sizes. t = u + v.}

\item{.precision}{Defines the precision by which confidence limits, p-values, and size is determined. Defaults to 1E-03.}
}
\description{
Calculate expected value of theta (OR) given m, n, t and precision.
}
\keyword{calculate}
\keyword{expected}
\keyword{odds}
\keyword{ratio}
\keyword{theta}
\keyword{value}
