% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_test_frame.R
\name{construct_test_frame}
\alias{construct_test_frame}
\title{Construct testing frame and randomisation values ---------------------------}
\usage{
construct_test_frame(.odds_ratio, .m, .n, .alpha, .precision, .message = FALSE)
}
\arguments{
\item{.odds_ratio}{The null hypothesis odds ratio being tested. No default.}

\item{.m}{Integer input responses and sample sizes. Tests u/m versus v/n. No default.}

\item{.n}{Integer input responses and sample sizes. Tests u/m versus v/n. No default.}

\item{.alpha}{The nominal significance level α. Defaults to 0.05.}

\item{.precision}{Defines the precision by which confidence limits, p-values, and size is determined. Defaults to 1E-03.}

\item{.message}{A logical. Defaults to FALSE. Setting this to TRUE will print messages as the function is running; this can be useful for debugging.}
}
\description{
Construct data frame with possible values for u and attach the gamma1 and
}
\section{gamma2 randomisation values (given the null odds ratio, precision, and}{
significance level alpha) to this data frame.
}

\keyword{construct}
\keyword{gamma}
\keyword{randomisation}
\keyword{testing}
\keyword{values}
