% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_gamma12.R
\name{find_gamma}
\alias{find_gamma}
\title{Find randomisation values gamma1 and gamma2 --------------------------------}
\usage{
find_gamma(c1, c2, .odds_ratio, .m, .n, .t, .alpha, .precision)
}
\arguments{
\item{c1}{Critical value 1.}

\item{c2}{Critical value 2.}

\item{.odds_ratio}{The null hypothesis odds ratio being tested. No default.}

\item{.m}{Integer input responses and sample sizes. Tests u/m versus v/n. No default.}

\item{.n}{Integer input responses and sample sizes. Tests u/m versus v/n. No default.}

\item{.t}{Integer input responses and sample sizes. t = u + v.}

\item{.alpha}{The nominal significance level α. Defaults to 0.05.}

\item{.precision}{Defines the precision by which confidence limits, p-values, and size is determined. Defaults to 1E-03.}
}
\description{
Find the gamma1 and gamma2 randomisation values (given the null odds ratio,
precision, and critical values) so that the test size is alpha.
}
\keyword{construct}
\keyword{gamma}
\keyword{randomisation}
\keyword{testing}
\keyword{values}
