% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_size.R
\name{local_size_gradient}
\alias{local_size_gradient}
\title{Find the local size gradient of the MFET as a function of the null OR theta_0.}
\usage{
local_size_gradient(
  nuisance,
  .gamma0,
  .odds_ratio,
  .m,
  .n,
  .df,
  .alpha,
  .precision
)
}
\arguments{
\item{nuisance}{Nuisance parameter.}

\item{.gamma0}{Some randomisation probability gamma0.}

\item{.odds_ratio}{The null hypothesis odds ratio being tested. No default.}

\item{.m}{Integer input responses and sample sizes. Tests u/m versus v/n. No default.}

\item{.n}{Integer input responses and sample sizes. Tests u/m versus v/n. No default.}

\item{.df}{Testing frame (data frame) generated as part as construct_test_frame().}

\item{.alpha}{The nominal significance level α. Defaults to 0.05.}

\item{.precision}{Defines the precision by which confidence limits, p-values, and size is determined. Defaults to 1E-03.}
}
\description{
Find the local size gradient of the MFET as a function of the null OR theta_0.
}
\keyword{find}
\keyword{gradient}
\keyword{hypothesis}
\keyword{local}
\keyword{size}
\keyword{test}
