% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_fe_test.R
\name{mod_fe_test}
\alias{mod_fe_test}
\title{Define MFET ----------------------------------------------------------------}
\usage{
mod_fe_test(z, .df, .gamma0, .odds_ratio, .m, .n, .alpha, .precision)
}
\arguments{
\item{z}{Vector containing (u, t).}

\item{.df}{Testing frame (data frame) generated as part as construct_test_frame().}

\item{.gamma0}{Some randomisation probability gamma0.}

\item{.odds_ratio}{The null hypothesis odds ratio being tested. No default.}

\item{.m}{Integer input responses and sample sizes. Tests u/m versus v/n. No default.}

\item{.n}{Integer input responses and sample sizes. Tests u/m versus v/n. No default.}

\item{.alpha}{The nominal significance level α. Defaults to 0.05.}

\item{.precision}{Defines the precision by which confidence limits, p-values, and size is determined. Defaults to 1E-03.}
}
\description{
Define the non-randomised non-conservative MFET. This is a function whether
to reject or not based on some value gamma0. Requires n, m, alpha, df,
gamma1, gamma2 and 'gamma0'.
}
\keyword{conservative}
\keyword{exact}
\keyword{fisher}
\keyword{modified}
\keyword{non}
\keyword{randomised}
\keyword{randomized}
\keyword{test}
