% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modified_fisher.R
\name{modified_fisher_exact_test}
\alias{modified_fisher_exact_test}
\title{The non-conservative, non-randomised modified Fisher Exact Test.}
\usage{
modified_fisher_exact_test(
  u,
  m,
  v,
  n,
  odds_ratio,
  alpha = 0.05,
  precision = 0.001,
  message = FALSE,
  maze = 10,
  method = "zoom",
  zoom_iter = 6,
  superiority = FALSE,
  power = TRUE,
  power_at_pi1 = 0.5,
  power_at_pi2 = 0.75,
  conf_int = TRUE,
  pvalue = TRUE,
  local_size_data = FALSE
)
}
\arguments{
\item{u}{Integer input responses and sample sizes. Tests u/m versus v/n. No default.}

\item{m}{Integer input responses and sample sizes. Tests u/m versus v/n. No default.}

\item{v}{Integer input responses and sample sizes. Tests u/m versus v/n. No default.}

\item{n}{Integer input responses and sample sizes. Tests u/m versus v/n. No default.}

\item{odds_ratio}{The null hypothesis odds ratio being tested. No default.}

\item{alpha}{The nominal significance level alpha. Defaults to 0.05.}

\item{precision}{Defines the precision by which confidence limits, p-values, and size is determined. Defaults to 1E-03.}

\item{message}{A logical. Defaults to FALSE. Setting this to TRUE will print messages as the function is running; this can be useful for debugging.}

\item{maze}{Number of points at each iteration to select the nuisance parameter with maximum size from.}

\item{method}{Defines the numerical method used to find the optimum nuisance parameter (maximising actual size) of the test. The default is "zoom", with the second option being "trust" (this uses the trust() function from the trust region package).}

\item{zoom_iter}{Number of iterations to zoom in with (in the "zoom" method).}

\item{superiority}{A logical. Defaults to FALSE. Setting this to TRUE will calculate the power for testing superiority.}

\item{power}{A logical. Defaults to TRUE. Setting this to FALSE will skip the power calculation.}

\item{power_at_pi1}{Power for pi1 (the null hypothesis). Defaults to 0.5.}

\item{power_at_pi2}{Power for pi2 (the alternative hypothesis). Defaults to 0.75.}

\item{conf_int}{A logical. Defaults to TRUE. Setting this to FALSE will skip the (1-alpha, two-sided) confidence intervals calculations.}

\item{pvalue}{A logical. Defaults to TRUE. Setting this to FALSE will skip the (two-sided, test-based) p-value calculations.}

\item{local_size_data}{A logical. Defaults to FALSE. Setting this to TRUE will attach a 'local.size.data' data frame to the test results used in plotting the size of the test.}
}
\value{
Returns an \if{html}{\out{<htest>}} object
}
\description{
The non-conservative, non-randomised modified Fisher Exact Test.
}
\examples{
\dontrun{
# Example here
}
}
\keyword{conservative}
\keyword{exact}
\keyword{fisher}
\keyword{modified}
\keyword{non}
\keyword{randomised}
\keyword{randomized}
\keyword{test}
