% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimise_gamma0.R
\name{optimise_gamma0}
\alias{optimise_gamma0}
\title{Find optimal gamma0 MFET ----------------------------------------------------------------}
\usage{
optimise_gamma0(
  .odds_ratio,
  .m,
  .n,
  .alpha,
  .precision,
  .method,
  .maze,
  .zoom_iter
)
}
\arguments{
\item{.odds_ratio}{The null hypothesis odds ratio being tested. No default.}

\item{.m}{Integer input responses and sample sizes. Tests u/m versus v/n. No default.}

\item{.n}{Integer input responses and sample sizes. Tests u/m versus v/n. No default.}

\item{.alpha}{The nominal significance level α. Defaults to 0.05.}

\item{.precision}{Defines the precision by which confidence limits, p-values, and size is determined. Defaults to 1E-03.}

\item{.method}{Defines the numerical method used to find the optimum nuisance parameter (maximising actual size) of the test. The default is "zoom", with the second option being "trust" (this uses the trust() function from the trust region package).}

\item{.maze}{Number of points at each iteration to select the nuisance parameter with maximum size from.}

\item{.zoom_iter}{Number of iterations to zoom in with (in the "zoom" method).}
}
\description{
Find the optimal gamma0 by maximising the size with respect to alpha.
}
\keyword{alpha}
\keyword{exact}
\keyword{fisher}
\keyword{gamma0}
\keyword{modified}
\keyword{optimise}
\keyword{test}
