% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_fe_test.R
\name{random_fe_test}
\alias{random_fe_test}
\title{Define randomised conservative fisher exact test ---------------------------}
\usage{
random_fe_test(z, .df, .odds_ratio, .m, .n, .alpha, .precision)
}
\arguments{
\item{z}{Vector containing (u, t).}

\item{.df}{Testing frame (data frame) generated as part as construct_test_frame().}

\item{.odds_ratio}{The null hypothesis odds ratio being tested. No default.}

\item{.m}{Integer input responses and sample sizes. Tests u/m versus v/n. No default.}

\item{.n}{Integer input responses and sample sizes. Tests u/m versus v/n. No default.}

\item{.alpha}{The nominal significance level α. Defaults to 0.05.}

\item{.precision}{Defines the precision by which confidence limits, p-values, and size is determined. Defaults to 1E-03.}
}
\description{
This is a function of z = (u, t), and is an indicator function whether to
reject or not. Requires n, m, alpha, and df. Mainly for comparing with MFET.
}
\keyword{exact}
\keyword{fisher}
\keyword{randomised}
\keyword{randomized}
\keyword{test}
